Initial version of the UI state data class for the encode screen.
Keeping with the standard model-view-controller architecture, this
commit provides a data class representing the UI state of the
encoding screen.

As of right now, the encoding UI consists of: A display of how
many characters have been entered, a text area for the message to
encode, a text box for the tag(s) to apply, and a read-only display
of the encoded message (to be directly copied or shared
to other apps.) The data class stores those parts of the UI that
are dynamic in nature.