Beginnings of persistent settings
This is a big commit, sorry!

This implements necessary logic to persist user settings across
app death. Saving is done via an explicit button, as there are tons
of nasty race conditions if you try to write to disk while typing.

This commit also includes key generation. That was super easy,
actually!

Finally, the import/export buttons have been moved from the top
bar. This was not a super sensible location for them, and having
them there would complicate the state architecture (e.g. making
settings state not local to the settings view.)