Implement more structured navigation and ViewModel provision.
This is a pretty big commit :) Basically, nothing operationally has
changed, but all of the ViewModels are now accounted for through a
provider class (that defines Factory) and navigation is handled in
a more structured way (via an interface and singletons implementing
the interface.)

Finally, the top-level of the app introduced a view-state manager
class to keep all of the page/screen-specific parts of the view
(e.g. top bar actions, the title, whether the navigation bar should
be show, etc.) in one place. This could also be used to help make
the app more adapatable to different screen sizes, something we
currently do not do.