Implement a 'models' table for Butkus/Cadenas settings.
Profiles must store a reference to the model that is to be used
when encoding/decoding with that profile. While in the UI the
model must be displayed as its name, the fact that model fetching
is an entirely separate process (and the nature of relational
databases) requires that models have their own table and be linked
to the profiles table via foreign key.

This commit implements the components to interact with a models
table via Room APIs (Entity, DAO, repository.)