Implement a repository providing the Butkus/Cadenas operations.
Though it's a bit atypical, this repository provides indication
of whether or not Butkus/Cadenas is initialized, and functionality
to initialize the backend whenever the chosen profile changes.

Additionally, this repository provides suspend functions to
encode/decode via the Butkus/Cadenas object.

The use of DataStore in this same repository saves the ID of the
profile that's selected, which as mentioned above, is used to
initialize Butkus/Cadenas (and, of course, to control the UI of
the profile management screen.)