Refactor existing work, implement better navigation, add profiles.
This is a classic huge commit :)

First and foremost, the app is now organized around more explicit
navigation graphs. The root graph connects the "processing" screen
(where one can encode/decode messages) to the web of settings
screens.

Being on separate navigation graphs, the processing screens become
ephemeral with respect to modifying settings - if you leave the
processing screen to edit settings, when you return, all content
will have been reset. I almost see this as a feature - in a pinch,
you can quickly tap the cog menu to erase sensitive/private
messages from the UI state.

The processing screen is quite close to the previous architecture,
and isn't all that interesting. The one notable change is the
'unification' of the encode/decode screens - since tags are now
associated with profiles rather than individual messages, there
was no need for the extra UI related to tags. This allowed for
some code reuse, which is always nice.

Where things get interesting are the various settings screens.
At the top-level is a simple menu, allowing the user to select
between profile and model management. Currently, model
management does nothing in the UI - but profile management is
fully implemented.

When the user follows the profile management menu, they are
presented with a radio list of all the profiles they have
entered. Tapping a list item selects it, and triggers a
re-initialization of Butkus/Cadenas.

Also on this screen are an edit button (when a model has
been selected), and a large floating action to add
additional profiles. These two screens are quite similar -
the former presents a form similar to the old settings
screen, with save and delete buttons, and the latter is
identical, except it only has a save button.

There are still some things worth refactoring here. One
that has been bugging me particularly is the work being
done in Kotlin that ought to be done in SQL queries. I
wasn't quite able to get the queries I wanted to work
in a DAO, but I'll be thinking about this more (since
we really shouldn't need to make multiple queries to
pair up profiles with their associated model details,
e.g.)

And, while I think these commits do a decent job at
high-level overviews of what's going on, the code is
woefully undocumented right now. This will become a
priority ASAP, since the architecture finally feels as
if it is "coming together". That and some tests will
do this a lot of good!