Implement improved profile deletion.
Previously, profile deletion happened on the profile edit screen.
This was actually the opposite of how we want things to be - the
active profile should not be able to be deleted at all!

This solves a number of problems at once. First, previously, the
backend would still operate as if the deleted profile was
selected. This was problematic, as there was an ID reference to
a now-deleted profile. Additionally, it should not be possible
to delete the last remaining profile.

The new solution, to move the delete operation to the profiles
screen and make the selected profile not delete-able, solves
both of these problems.