Implement a model-management screen.
This commit adds a screen similar to the profile manager, but for
models. It is much simpler, featuring the same FAB, but no edit
functionality, and no selection mechanism. There is still a way
to delete items - all but the last model can be deleted, much
like with profiles.

When a model is removed, all associated profiles are also
removed. The user is warned of this behavior in the
confirmation dialogue.