Reorganize settings around models.
There is a one-to-many relationship between models and profiles.
It is more natural, then, for the UI to group profiles by model,
rather than altogether. This also simplifies the code, resulting
ultimately in fewer databse transactions since we now 'know' the
model ID when adding a new profile.

This commit also makes a number of cosmetic changes. The edit and
delete options for profiles are now in a "dots" menu, rather than
cluttering the ListItem and leaving little room for the profile
description. A similar change was made to the model management
screen, though rather than "Edit" there is a "Profiles" menu in
the dropdown.

It is now impossible for the user to delete, directly or
indirectly, the profile that is currently selected. This was made
possible by only enabling delete for those profiles which are not
selected, and those models that are not the parent of the selected
profile. Deleting an inactive model will remove all associated
profiles, maintaining database consistency. Activity of a model is
indicated by a check mark on the corresponding item on the model
management screen.

A natural improvement to UI/UX from here is the addition of a
"First Launch" sequence, which will walk the user through the
creation of an initial model/profile. This will guarantee there is
always at least one configuration profile for Butkus/Cadenas to
use.