Simplify model representation, remove model editing, name refactor.
First and foremost, this commit removes the ability to edit models
that have been downloaded. Since the model name is used as a
directory in internal storage and there is no good reason to alter
the URL post-download, the inclusion of an edit screen is
extraneous.

This also implied that model descriptions are a bit problematic -
they ARE something that should be editable, even for something as
simple as a typo. Rather than leaving the world's most boring edit
screen, we opt to do away with model descriptions; they weren't
really adding anything to the UX anyway.

Finally, we renamed "ConfigDao" to "SavedConfigDao" to match the
underlying data class / be more consistent with other naming
conventions used in the project.