Implement model downloading action.
This commit implements the action of the 'Download' button on the
"Add Model" screen. The ConfigRepository exposes a Flow giving the
WorkInfo for the model-downloading worker, as well as a method to
enqueue the job.

The ModelAddScreen uses this state to create a LaunchedEffect that
listens for changes to the WorkInfo.State. This largely controls
the display of snackbar messages and a progress indicator, as well
as writing to the config database once/if the download completed
successfully. It would be nice if, instead of keeping the user on
the download screen, the snackbar messages could be displayed
anywhere and the user can do other work while the download takes
place in the background. This should merely require lifting the
SnackbarHostState to the top-level of the app, and feeding it to
all screens utilizing Scaffold.

A number of data model/architectural improvements to be made became
apparent while implementing this functionality. Namely, the subtle
duplication of data between the filesystem and configuration
database leaves lots of room for data integrity to fall apart. It
can be the case that the models on the filesystem are not reflected
in the database, for instance - the file-downloading job can survive
app death, but updating the configuration database cannot. This is
problematic and effectively means that we have multiple sources of
truth for downloaded models.

The solution to this problem is to flatten the configuration DB,
and only rely on the filesystem for model information. This, funny
enough, means that the app flow will more resemble a previous
iteration, where model/profile management were separate. Since we
do need to save model URLs for (eventual) config sharing, we can
just write an additional .metadata file to model directories with
this information, to be read when needed.