Cadenas.kt: Make encode/decode non-suspending functions.
This commit moves the delegation of encoding/decoding work to
threads into the ViewModel code, which makes more logical sense
than the previous arrangement, where the encoding process was
explicitly returned from a coroutineScope.

This simplifies the business logic to just be the core operations
and makes it clear that the ViewModel will launch a suspending
block on CPU threads for these expensive operations.