settings.gradle: Enable build caching with Artifactory.
This commit configures the Gradle build to use remote build
caching via Artifactory. This is primarily to support CI,
but will benefit anyone forking the project if they, too
have Artifactory set up.

This commit does _not_ include instructions in the README
to configure your own Artifactory credentials - that will come
later. It's worth exploring if it's possible to conditionally
use local/remote based on the presence/absence of credentials
in the Gradle properties file.