settings.gradle: Use a composite build structure.
This commit updates the app to use a Gradle composite build rather
than a multi-project build. This builds cadenas-core independently,
using the compiled library directly (see build.gradle).

This resolves any issues related to there being two settings.gradle
files (for cadenas-core and the app), and will "just work" if the
core libraries are ever independently uploaded to a package repo
and imported as any other package (i.e. we don't need to rely on
having the source code for the core API.)