CadenasApplication.kt: Create a notification channel.
If the app is built for SDK >= 26, create a notification channel to
be used by the application.

Note that as of this time, this is a redundant check, since the
application has a minimum SDK of 26 (Android 8.) The check is
included in case we wish to support even older versions of Android
in the future.