Configure signed release builds.
This commit adds signing back to the release-build configuration,
and documents the change in the README so maintainers can configure
their local environment to easily produce release-mode APKs.

CI will use this configuration on release-mode jobs (i.e. for the
main branch) to make official APKs easily available as artifacts.