Use lower-API-compatibile notification builders.
Notification.Builder is not available for API levels below 26 -
for this, we need NotificationCompat.Builder. This commit makes
this change.

Note that this does _not_ change the application to be compatible
with an API level below 26. This would require a substantial set
of changes, namely to avoid the use of the Kotlin standard
kotlin.io.path package, which requires API 26.