Add text intent filter to AndroidManifest.
To support the receipt of text from other applications, Cadenas
must define an intent filter for SEND actions of text/plain MIME
data. This will show Cadenas in the sharesheet for such data,
allowing the activity to be launched.

This is merely the first step towards allowing text to be shared
to the application, and doesn't 'do' anything with the received
data.