Remove explicit navigation from final setup screen.
There was an awkward visual artifact in the app, causing the
setup screen to be briefly visible when starting the app after
completing the first-time setup.

This commit starts to correct this behavior, by first eliminating
the explicit navigation from the final setup screen into the
processing screens. This 'navigation' can instead be handled by
the shared preference that tracks whether first-time setup has been
completed or not.