Simplify the first-time setup check.
This completes the work of the previous commit, simplifying the
root navigation host (to exclude the first-time setup) and moving
the logic to choose between first-time setup and regular operation
to the top-level of the application (i.e. the top-level Composable)

This greatly simplifies much of the navigation graph, and removes
the visual artifact described in the previous commit message.