Simplify the processing screen and finish sharing implementation.
Previously, the processing functionality of the application was
split between two screens, navigated using a bottom navigation
bar.

This commit removes this navigation, instead opting for a mode-
based approach: The user selects between encoding and decoding
mode, inputs text, and literally hits 'Go!'.

This simplifies the implementation of deep-linking, as now the
result of SEND text/plain intents is simply (1) selecting the
proper mode (decode) and (2) populating the input box.

This commit also contains all of the supporting code for this
change, namely the addition of an enum class representing the
mode of operation, an overhaul of the associated ViewModel,
and restructuring of the UI to eliminate the NavHost.

As a somewhat nice side-effect, we no longer need the state-
holder class to handle screen titles and sharing behavior;
this can all be controlled directly by the UI state now that
it is aware of the processing mode.