Profile.kt: Add extension method to convert to QRCode.
This commit provides a method that converts Profiles to QRCodes.
In particular, the encryption key, model prompt, and model name
are joined with '-' to create the QRCode (a format that is easy to
parse when we later implement the scanning of QR codes to import
channel information).

Note that the use of the model name is technically undesirable,
in both this prototype version and the anticipated alpha release
version where model management is significantly different. We
ought to use a hash of the model files for comparison purposes
rather than the name (as the name will be determined by how the
model is added to the user's device and will not uniquely identify
the model).