ProfileRepository: Add method to save QR image for a profile.
This commit adds the saveQRForProfile method to ProfileRepository.

This method saves the QR code corresponding to a Profile as an
image on the user's device via the MediaStore API, following
recommended practices for Android versions below 10.

This functionality backs the 'Save' button that will be on the
profile export screen.