More idiomatic use of repository and view model methods.
This commit does some refactoring of the profile export screen and
associated functionality, including:

- Moving the conversion of QRCode to byte array and bitmaps to
  extension methods defined on QRCode.
- Simplifying the view model to only store the bitmap as state.
- Hiding the image container (and disabling the save button) when
  the image has not been created yet.
- More consistent method naming and documentation.

The most important effect of these changes is the experience using
the app: Before, the export screen would 'jump' into view as
creating the bitmap blocked the main/UI thread. Now, the transition
is immediate, and there may be a moment before the image appears.