Implement the profile importing screen.
This commit implements profile importing screen, which allows for
users to scan QR codes to import the key material associated with
a Cadenas channel.

First: The method to parse Profiles from Strings has been moved,
and redefined as an extension method on String.

The screen itself consists of an Android Camera preview running
our QRAnalyzer. When a valid Cadenas QR is scanned, a dialog
allows the user to either continue (bringing them to a screen to
set the channel name and description) or reset and scan again.