Use Column instead of LazyColumn for profiles and models.
As noted in an earlier commit, the profile-management screen was
not functioning properly. At some point during screen composition,
the LazyColumn used to display the list of profiles was attempting
to index past the end of a collection whenever a profile was
deleted.

This commit solves this issue by using a regular Column modified
to be scrollable instead of a LazyColumn. This works, and will
suffice for our use-case (as the number of profiles or models
will typically be quite small compared to e.g. conversations in
a messaging application).