Remove explicit encryption key generation from the UI.
Previously, Cadenas channels created from scratch had their keys
explicitly generated by the user in the UI. This required users to
have to interact directly with the idea of an encryption key, which
is undesirable from a usability perspective.

Now, key generation happens automatically when a new channel is
being created. This works in tandem with the QR code functionality,
which enables these secret keys to be shared without ever seeing
them explicitly.