Remove the settingsRepository and singleton Cadenas object.
This is a relatively large commit spanning much of the app.

First, the SettingsRepository class has been completely removed.
Since it was only necessary in a setting where the TextCover
needed to be initialized at app startup, it was prudent to remove
now that channels load their associated models as needed.

Next, the ChannelRepository is given a method to create a TextCover
from a channel ID. This eliminated the need for the Cadenas class,
so it has been removed as well.

To reflect the new navigation graph, a number of packages have been
moved (namely, the channels package (with add/edit/import/export
screens) has moved to the top-level of ui alongside the home screen,
processing, and settings). This change is also reflected in the
navigation graphs themselves.

On to a downside: With this new structure, sharing to the app
directly does not make sense (as the processing screens are unique
to the channel being used to do the processing). There would be no
way to pick one such screen from a text/plain intent, so this commit
also removes all code associated with deep-linking to processing on
these share intents.

Model-management is dramatically simplified in the wake of the
settings repository being removed. Like with channels, it is now
possible to delete models at any time.

The settings screen now only has one navigation target, namely for
model-management. Eventually this may be expanded with preferences
or other settings-adjacent concepts.