Remove hashtags from the app.
Previously, Cadenas supported channels having an associated
hashtag. This was useful for ROCKY/RACE to sort messages in a
high-volume setting.

Since that is not the use-case for the technology now, and since
the idea was reportedly arcane/difficult to understand, this commit
removes the concept from the app entirely.