Use a FAB instead of top-bar actions on the home screen.
This commit moves the top-bar actions of the home screen to a
dropdown menu opened by a floating action button (FAB).

This better emphasizes the main action on the screen.

Note that an otherwise empty bottom bar is used to make sure the
FAB never covers up one of the channel items (an issue we have had
in the past).