ModelDownloadWorker: Add more robust error-reporting.
First, a minor change that has no observable effects: We now
subclass Worker instead of CoroutineWorker, since we have no need
for coroutines in this operation - the work is already being done
in a separate thread.

Previously, the model-downloading worker did not provide very good
responses to the user upon failure, and indeed, some cases would
fail without reporting anything at all.

This commit rectifies this sorry state of affairs, adding checks
for (in theory) all possible exceptions that may occur when trying
to download a model from the Internet. In particular:

- We fail if an exception occurs while setting up a connection in
  the first place. It is unlikely that this will ever occur, as
  the UI prevents invalid HTTPS URLs from being entered.
- We fail if actually establishing a connection fails. This check
  covers cases where, for instance, a syntactically valid but
  nonsense HTTPS URL is entered. In the special case of a timeout,
  we appropriately indicate that is what occurred.
- We fail if the content-type header is not application/zip. This
  may be too conservative a check, since it is permissible for the
  content-type header to be absent in an HTTP(S) record. We can
  easily remove this check if it proves too aggressive for client
  needs.
- We fail if an input stream cannot be created from the established
  connection.
- We fail if there are issues decoding the ZIP.