ModelAddScreen: Dramatically simplify state management.
This commit primarily reduces the complexity of the UI state
controlled by the model-downloading worker.

First, there was a serious redundancy: A Boolean state was
subtly duplicated in a way that made understanding the state of
the whole system difficult.

Furthermore, there was no real need to pattern-match on the worker
state, as the behavior in one of the two branches was redundant
given the above de-duplication of Boolean state. As this left only
one meaningful branch, a simple conditional statement suffices,
which dramatically reduces the nesting of the code.

Finally, there was an extremely subtle bug where the screen
could enter a state where the progress bar was showing (and
therefore all the buttons were disabled), but the background
worker had already failed. It turns out that it was possible
for the WorkerInfo.state to go unchanged between button presses,
meaning there was a 'missing' launch of the effect that shows
a snackbar message. To correct this, this commit keys the
launched effect on the entire WorkerInfo rather than just the
state value.

Beyond these significant changes, this commit also includes some
changes to the names of local variables, reordering of UI effects,
and a more polished version of progress indication - in particular,
we no longer display "Extracting ..." for a brief time before the
work starts. The progress bar itself still shows up briefly in
error cases, but this seems like a reasonable UI behavior.