ModelAddScreen: Polish the display of the progress bar.
This commit is a reflection on part of a previous commit in which
I said that it seemed reasonable to show the progress bar briefly
even in error conditions.

This doesn't really make sense if it's unreasonable to show the
text part :) So, this commit goes back on that previous decision,
making it so the text _and_ progress bar are gated by the filename.