ModelDownloadWorker: Compute a hash of the model weights.
This commit uses a DigestInputStream in the model-download worker
to compute an MD5 hash of the model weights, to be stored in the
database and eventually used for QR code / channel-sharing.

MD5 is sufficient/acceptable, since we are just using it as a
checksum. There are no cryptographic requirements on these hashes.