Model.kt: Remove the integer primary key, using the name instead.
This commit removes the unnecessary integer primary key from the
Model entity class, and uses the name instead.

This is a justifiable primary key, as model names correspond to
distinc locations in the filesystem. This will be enforced at the
UI level by rejecting model names that are already in use.