CadenasDatabase: Relate the model and channel tables.
This commit relates the model and channel tables via ForeignKey,
and leverages this relation to simplify the cascading deletion of
channels when the model they use is deleted.

This simplifies the model-management use-case to a point where it
is now a thin wrapper around the model repository. This will be
addressed in a future commit.