ManageModels: Use the database to list models and simplify deletion.
This commit serves two main purposes: Further simplfiying the
model-management use-case (with the goal of eventual deletion),
and reverting to the use of the Room-generated delete method for
the model table.

This latter decision was made for consistency with the home
screen implementation, and the fact that the previously-stated
benefits were not quite benefits :) Neutral, at best.