Correct bug where channel add screen could not show models.
Previously, there was a bug where the channel-adding screen did
not correctly show the dropdown of available models.

This was due to a subtle recomposition-keying bug, where the
callback controlling the opening/closing of the dropdown was not
being reconstructed with an up-to-date value for the models list.

This commit fixes this problem by keying recomposition of the
dropdown on the model list.