Preliminary implementation of proper QR importing with model hashes.
This commit implements the proper check on QR codes when importing
to do so with a model hash - see the implementation of
String.toChannel for the 'meat' of this check. Briefly: We add a
method and piece of state for the "model in a QR" to the
view model that queries the database for a model with the given
hash.

While this is functioning properly, the UI is too opaque when a
valid QR is found but the model hash is not present on the device;
in this commit, nothing happens at all in that case.

A future commit will improve the error messaging here.