Add more robust handling of valid QRs but missing models.
Previously, Cadenas would simply ignore QR codes that contained
hashes for models not installed to the device, even if the QR was
a valid Cadenas channel barcode.

This commit adds more robust checking, reporting an error if a
model with the hash has not been downloaded and immediately
navigating to the model-downloading screen.

There is an interesting but inconsequential visual bug where,
briefly, this error alert will show even for a channel that _can_
be added from the barcode. This is slightly ugly, but does not
affect the user experience as the proper dialog almost immediately
takes its place. More investigation will be required if we really
care about smoothing out this only slightly rough edge.