Add an option to add a model to the dropdown when creating channels.
This commit implements AlexM's excellent idea to allow the user to
navigate to the model-adding screen directly from the new channel
creation screen, if the model they want is not in the dropdown
list.

This has the added benefit of removing the subtle composition bug
addressed in an earlier commit, and no longer requiring the use of
an explicitly-keyed composition.