Use consistent navigation parameter naming and backstack pops.
This commit finally makes the naming of parameters related to
navigation consistent across the codebase, and additionally removes
the pre-existing separate notions of navigating up and navigating
back. For our purposes, navigating back via backstack popping is
sufficient (and where it "isn't", we can manually specify the
navigation target).

This change provides a more intuitive experience when, eg, adding
a model from the new-channel screen, since the navigation will now
take the user back to the new-channel screen (rather than to the
list of all models). This principle has been applied to all of the
screens in the application, except the home screen which is the
one "special case" where navigation goes to the settings screen.