build.gradle: Add Google shortcuts integration dependency.
To more easily handle the dynamic shortcuts associated with
Cadenas, this commit adds a dependency on the Google shortcuts
integration library.

This required the addition of an exclusion rule for the Google Tink
library, as the shortcuts library depends on a different version
than cadenas-core; without this rule, the build fails with many
"Duplicate class" errors for Tink classes.