HomeNavigation: Deep link to the home screen rather than processing.
There is unfortunately no way to get direct sharing to work the
way we really would like it to (ie, treat Cadenas channels like
contacts and provide a sharing interface to match).

Instead, we will link to the home screen, adjusting the title to
indicate that the chosen channel will be used for decoding of the
text shared in the received intent. This commit simply moves the
deep linking to reflect this.