TextCover: Improve the memory usage when encoding/decoding.
See the recent commits on cadenas-core for an explanation of the
memory leaks in PyTorch.

This commit changes the initialization of the TextCover used to
encode and decode messages to occur when those operations are
invoked. Furthermore, it uses the newly-added explicity destroy
method on TextCover to free up the PyTorch memory when the
operation is complete.

Profiling the application confirms this solves the earlier
memory issues with the app, where nearly a gigabyte of RAM was
being held on to.