AppContainer: Use interface instead of concrete class for channels.
The previous commits renamed the original channel repository class,
and introduced an interface to enable proper dependency injection (and
better testability of upstream / dependent classes and objects.

This commits completes the transition by reverting the changes to
parameter names introduced by the renaming; in stages, we've
effectively inserted a layer of abstraction without damaging the
structure of code already relying on the interface being introduced.