strings: Remove 'Go' and add new strings for processing.
Previously, the ProcessingMode type was used to determine the string
labels for the processing screen upon mode switch. This left the
strings untranslateable, so this commit adds string resources for
these modes and uses them in place of the ProcessingMode toString
conversion.

This commit also replaces the string 'Go!' with 'Execute', which has a
more obvious translation in Spanish.