ModelAddScreen: Add a handler for the sample-download button.
This commit replaces the stub onClick for the sample-model-downloading
button with an acutal implementation.

The only difference from the 'normal' download button handler is a
trick to make sure the sample model also gets properly saved to the
app's database; in particular, the trick updates the screen's UI with
the name and URL of the sample model, as these UI elements are used to
determine the contents of the record stored in the app's database.