ProcessingScreen: Add a warning to the user about editing covertext.
It is crucial that output covertexts are not modified before being
shared; the correct operation of Cadenas depends on reliable
communication of the covertext bytes.

This commit adds an alert dialog warning the user about editing the
output, that is shown after an encoding job is complete / before the
user has a chance to copy or share the output.