Added dialog after successfully saving a channel QR code to the photo library.
As I didn't want to change the onNavigateBack functionality, I added a new onNavigateBackAfterSave to the ChannelExport Screen.  OnNavigateBackAfterSave navigates to the home screen but passes a new saveSuccess boolean argument (see CadenasRootNavGraph).

HomeNavigarion processes the boolean argument and passes it to the HomeScreen.

The home screen tracks the saved boolean and will display the Snackbar as a launch effect.