build.gradle: Disable linting for release builds.
(See
https://stackoverflow.com/questions/73415617/composite-builds-broken-on-android-or-did-i-do-this-wrong)

The release-mode lint check repeatedly failed with an error related to
the composite build structure of the project (essentially, the linting
checks could not find the cadenas-core JAR). The above SO post details
a similar behavior, and a comment from May of this year suggests that
this is a known problem with composite builds for Android.

This commit disables the lint check in release builds, to avoid this
problem (particularly in CI, which handles our APK releases).