ModelAddDiskScreen: Implement the screen to add models from disk.
This screen is very similar to the existing model-adding screen, that
allows users to download pre-packaged models from the Internet.

It provides an alternative installation method, from the device's
media store via Android's content provider system. A button activates
a ZIP picker, a name is provided, and an installation button activates
a background process to unzip/copy the model files to the app's
resources so PyTorch can access them.