ModelAddDisk: Make the file a String? instead of a String.
Previously, even if the user navigated back from the file-selection
dialog without choosing a file, the 'Install' button would activate
and the label on the ZIP-choosing button would reflect that a ZIP had
been chosen, even though it hadn't.

This commit fixes this behavior by updating the file String to be a
String?, properly handling the null case on the frontend.